requirements.txt
# Heroky Package -------------------
django-heroku
gunicorn
whitenoise

# Django Packages -------------------
Django>=3.1.7
dj-database-url
dj-email-url
django-cors-headers
django-mathfilters
django-tables2
djangorestframework
django-widget-tweaks

# Other Packages --------------------
CairoSVG
defusedxml
mysqlclient
openpyxl
pandas
pycparser
python-dotenv
WeasyPrint
webencodings
xlrd==1.2.0
